Move <script> elements to bottom of <body>
Don't allow javascript to hold up page rendering. Peformance was
bitterly slow on the search results page prior to this change
(thanks to the availability status).

Of course, this change broke the availability status script (both on
results and item pages) but that's further proof that it needed to
be overhauled. That's next on my list.