DNA #282: Update handling of online access
Work in progress. Add "Online" button next to item record but the URL
is not returned by the availability service. If we're okay with
keeping the Online button at the very top or bottom of the list I can
get this working now without modifications to the availability
service. Otherwise, I'll need to call in Rick or John for backup.