DNA #282: Remove duplicate online availability
Getting around the fact that we don't have the online access URLs
returned from the availability service by placing it at the top. Call
number and notes for this item record will appear below the online
link.

Had to move the Availability <h4> to _show_holdings partial when it
should be with the holdings partial. If the UX team is okay with this
I'll convert the holdings partial to ERB (currently HAML).

Holding styles updated to give less weight to notes.