DNA #365 Selected items is functional again on dev
Now let's see what happens in production.

Turns out this was due to upgrading the jquery rails gem from 2.1.4 to
2.2. For now we'll just specify 2.1.4 in the Gemfile.

When using the 2.2 gem an error is thrown in response of the ajax
call on line 88 of checkbox_submit.js in the blacklight gem.
https://github.com/projectblacklight/blacklight/blob/master/app/assets/javascripts/blacklight/checkbox_submit.js

We get a 200 HTTP response (OK), but there is a parsing error:
"SyntaxError: Unexpected end of input"

You can see this for yourself by changing up lines 93-94 of said file:
  error: function(xhr, textStatus, errorThrown) {
    alert(xhr.status + ' | ' + textStatus + ' |' + errorThrown);