DNA #437: Display some_available as available
Previously any holding record that was only partially available
(multiple copies with at least one of the copies unavailable, but also
at least one available) was being displayed with the icon-time on the
item view. This was confusing since in the search results, the status
for this same item was displayed as available with icon-ok. Now the
availability status is consistent in both views.