DNA #324 Add location specific map/hours links
From the availability well on the item view, link to
http://www.library.cornell.edu/libraryhours with correct 'loc' param
value for the library location of holding in question.

Use hash map of substrings for holdings service location codes to
param values for the CUL site. Since holdings service location codes
have overlap for Annex (anx), it's first in the list so it trumps any
secondary location information in the code. Need to verify with team
that this is desired behavior even for RMC (rmc).

* this results in Kroch RMC (request in advance) linking to Annex
  instead of RMC