DNA #584 Group holdings into circulating & rare
This is a work in progress. The basics are functional now so Jenn can
take it further. Holdings for the item view are grouped into
circulating and rare (with rare sorted second) for display purposes. I
want to move much of this new logic into a helper method, to keep the
view clean.