Remove CLIO's unused add_display_elements helper
This method was mostly gutted and commented out from it's original
copied-from state in CLIO and the two remaining things it did do were
unecessary (location name is already provided in the hash and we
aren't using images for status).

In the process of this cleanup I stumbled across some major
duplication in _reqho.html.haml and _reqho_short.html.haml --
basically carbon copies of an outdated version of holdings.html.haml.
We need to DRY this up by removing both of these partials and updating
holdings.html.haml so it can address the needs for Aeon intermediary
screen.