Further tweaking of radio button value as array
Continutation of f04b5ca24.

John, I went with index-1 for the increment so that the first
boolean_row is 1 (second row) and the second is 2 (third row), etc. I
thought this would make it easier for you by keeping the index aligned
with the rest of the arrays (1st row = zero), but please let me know
if you need this changed.