DNA #652 Remove Mash class
Part of the upgrade from Blacklight 4.0.1 -> 4.2.0 (3d8d12c54)

We had mash.rb in our local app, but running a diff on our version vs.
what was in the core prior to it being removed/deprecated shows the
only changes were in white space. This file was added by Cline as part
of advanced search in 502583cef but again, I can find no evidence that
he made any changes to this file so I'm not sure why we need a copy
locally in our app. Some initial testing of advanced search after
removing the local copy of this file reveals no errors.

If it turns out that we do need a local copy of this file, we should
revert this commit and then adjust our local copy to reflect the
changes in core (deprecation of Mash class).

https://github.com/projectblacklight/blacklight/commit/804d38819f86a89052d1e0b52b7697020f09f846