DNA #652 Update path to mash.rb in solr_helper.rb
Part of the upgrade from Blacklight 4.0.1 -> 4.2.0 (3d8d12c54)

We had a copy of mash.rb in our local app, but I couldn't figure out
why so I removed it in a previous commit (c6ed1a5b0).

It turns out that we needed it becuase we also have a copy of
solr_helper.rb (Cline brought it in for advanced search) in our local
app which requires the Mash class. I've now updated that require to
use the path to the Blacklight gem so we can minimize the number of
core Blacklight files that are copied to the app -- trying to keep the
pain and suffering at a minimum for maintenance/upgrades.

https://github.com/projectblacklight/blacklight/commit/804d38819f86a89052d1e0b52b7697020f09f846