DNA #531 Drop batching from holdings request
Evaluate performance hit when retrieving availability from holdings
service for all search results on first page in a single request.
Testing page load time using Chrome's Speed Tracer found averages of:

* 6.3 seconds for 20 results

* 12.2 seconds for 50 results

Regardless of whether calls to holdings service are batched or sent in
single request, the search results were rendered in the browser within
3 seconds.

But the main issue that this effort was attempting to address is that
while the browser is working with the XMLHTTPRequest to the holdings
service, it will not respond to additional user action that involves
another AJAX request. Examples of such user actions include:

* clicking the "See more >>" link for any facet

* selecting/de-selecting an item in the search results

Browsers limit the number of concurrent requests to any single domain
name and the theory was that if we reduce the availability checking to
a single request to the holding service, that would free up a request
slot and allow for the additional AJAX requests listed above to fire
immmediately. Unfortunately, we have not found this to be the case.

In addition to this we also registered domain names for the holding
service (holdings.library and culholdingsdev.library), but this had no
effect on minimizing concurrent requests to the primary domain where
the Blacklight app is running since the availability requests are
vetted through an internal call to the app instead of a direct request
to the holding service. This is done so that we have a controller ->
view interaction and can use the view to parse the JSON response
instead of handling this in the JS.

As an example, the internal call to the backend controller:
/backend/holdings_shorthm/8002892/8003528...

Results in this request to the holdings service (dev domain used):
http://culholdingsdev.library.cornell.edu/holdings/retrieve_detail_short/8002892/8003528...

So what's next? Some options include:

* move the parsing logic out of the holdings_short.html.haml view and
  into holdings.js.coffee so we can make our AJAX call directly to the
  external domain

* maybe try moving the parsing logic that's currently in the
  holdings_short.html.haml view into the holdings service (a separate
  rails app) so that the json returned can just be injected into the
  page with minimal parsing required in JS