DNA #708 Word wrap fix for IE8+ (long URLs)
Thought we had this squashed with Jenn's fix in b45bec43a but perhaps
IE8 slipped through the cracks previously. Luckily, this was a quick
fix thanks to the notes I left myself on 1447c0fa7 in the itemview
branch. That led me to review the css-tricks post once again:

http://css-tricks.com/snippets/css/prevent-long-urls-from-breaking-out-of-container