Remove unused JS originally copied from CLIO
blacklight_application.js has been around since 12/3/12, which is the
date when the blacklight-cornell repo was created. It was introduced
as part of the 4th commit in the log (0b28a00a9) and hasn't been touched since:
https://git.library.cornell.edu/blacklight-cornell/commits/dev/app/assets/javascripts/blacklight_application.js

Due to issues originally setting up the git project, we don't have
full history for all of our files, but this one was most likely copied
from CLIO, Columbia's Blacklight instance. You can see that this file
was recently removed from the CLIO project as well:
https://github.com/cul/clio-spectrum/commit/9293ac28743d3add51e243ae79f9b40150a52dfe

Wherever it came from, I did verify that we're not using either of the
functions defined in this file at the time of removal.