DNA #742 Fix regex for display of date coverage
Previously, I was limiting to 7 trailing characters after the "to "
because I thought the only two possible values were:

* mm/yyyy

* present

(both 7 characters in length). So I thought I was playing it
safe but it turns out there are other values such as "3 months ago"
like here:
http://search.library.cornell.edu/catalog/5383976

So really, I was just playing it stupid.