DNA #726 Render WCL results only if keyword exists
So we're only offering to expand search results if the q param is
present. This is a continuation of the initial attempt in 5fecc4ac0
which was causing errors due to the expansion box trying to render
when for queries without q param contstraints (searches with facets
but no keyword).

No need to override query_has_constraints? helper now that checking
for the presence of params[:q] appears to be doing the job.