DNA #751 Add reliable indicator of search type
I had set up the expand your search function to only trigger if we
have a search term and we're not advanced searching.

I was using the :search_field param as the indicator for search type
but this is unreliable because John will switch this from the expected
'advanced' to some other value if the user submits an advanced search
with every row using the same field. See his bozocheck on line 360 of
cornell_catalog.rb for more details.