DNA #767 Click-to-search is working again
Search expansion (specifically the WCL Search API) does not like the
click-to-search queries, so we need to add click-to-search to the
growing list of condition checking for when we can expand a search
query.

I added a new helper method for this (expandable_search?) and did some
reorganizing of helpers so it was available to both the controllers
and the views.