DNA #799 Ensure GA tracking code injects all pages
Previously, following the installation instructions for the
blacklight_google_analytics gem, the extra head content was only
injected via the CatalogController. This is taken care of by the gem's
generator:
https://github.com/jronallo/blacklight_google_analytics

As a result, only catalog views had the Google Analytics tracking
code. By moving the include
BlacklightGoogleAnalytics::ControllerExtraHead to the
ApplicationsController, mission accomplished.