DNA #755 #801 Expand search w/ bento_search gem
Article count is now included since bento_search gem utilizes Summon
API. There's still some refactoring work required for the generic
links (#801) to WCL and Summon when a keyword is not present.

This commit introduces a new configuration file
(config/search_apis.yml) for the sensitive data related to API
connections. This file is not tracked, but an example is provided
(config/search_apis.yml.example).