DNA #752 fulltitle_vern_display in results for RTL
In place of the separate title_vern_display and subtitle_vern_display.

fulltitle_vern_display is currently handled by the '*_display'
dynamicField which means it's multiValued, unlike the rest of the
display fields used for the title. The application code could be
cleaned up a bit if the schema is updated so this new field is also
single valued.