DNA #695 Use OCLC # for cover img via Google Books
Almost 96% of our collection has an OCLC #, compared to less than 35%
with an ISBN. Overall we have more cover images using OCLC #, but I've
found that there are some items that had covers via ISBN which are now
missing, so we might want to improve on this further by using OCLC #
first and then ISBN for any potential stragglers. I'll add another
subtask as a potential future improvement.