DISCOVERYACCESS-1799 Status and/or location is different on search results vs. item view: Resolve this problem
when there were no item records, then the holding abbreviation was used to display the location, rather than the "display" string for the
location, which can be vastly different.

I also updated some tests to make them less worthwhile because I do not know what results are right -- the ones from the bento search, or from the catalog search,
and the counts differed.


Former-commit-id: 4bab08ca8704f547a1a37e5194d0e8481138fcbd